CLIENT-GAME: Modif now menu can init entity with the singletone MenuFactory MINOR
