+++ src/Client/Systems/Menu/ButtonCallbacks.hpp
@@ -1,0 +1,16 @@
+/*
+** EPITECH PROJECT, 2023
+** R-Bus
+** File description:
+** ButtonCallbacks
+*/
+
+#pragma once
+
+namespace Menu {
+    namespace Callback {
+        void initConnection();
+
+        void defaultCallBack();
+    }
+}
