+++ src/Client/Systems/Menu/CMakeLists.txt
@@ -10,4 +10,6 @@ target_sources(
     ${PROJECT_NAME_CLIENT}
     PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/Menu.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/MenuSystems.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/ButtonCallbacks.cpp
 )
